+++ printguard/schemas/__init__.py
@@ -1,0 +1,6 @@
+from schemas.schemas_cameras import *
+from schemas.schemas_events import *
+from schemas.schemas_networking import *
+from schemas.schemas_printers import *
+from schemas.schemas_settings import *
+from schemas.schemas_setup import *
