+++ printguard/utils/camera_utils.py
@@ -7,7 +7,7 @@ import glob
 
 import cv2
 
-from ..models import CameraState
+from schemas import CameraState
 from .camera_state_manager import get_camera_state_manager
