+++ printguard/utils/config.py
@@ -15,7 +15,7 @@ import torch
 from platformdirs import user_data_dir
 
 from .model_downloader import get_model_downloader
-from ..models import AlertAction, SavedKey, SavedConfig
+from schemas import (AlertAction, SavedKey, SavedConfig)
 
 # Config version - increment this when the config structure changes
 CONFIG_VERSION = "1.0.0"
